requirements.txt
pyrotgfork
tgcrypto
motor
python-dotenv
aiohttp
asyncio
APScheduler
pytz
google-generativeai
deep-translator
openai